[benchmark] Reorganize C++/Foundation benchmarks into separate files
